move files tab context menu handling in ui component moved actions component to file
